добавил не все варианты, так как на мой взгляд тут далеко не везде оно надо. Ну и чтобы читаемость не упала

+++ methods.py
@@ -21,20 +21,14 @@ def create_folder():
 def remove_folder():
     name = input('Введите название элемента: ')
     if os.path.exists(name):
-        if __is_file(name):
-            os.remove(name)
-        if __is_folder(name):
-            os.rmdir(name)
+        os.remove(name) if __is_file(name) else os.rmdir(name)
 
 
 def copy_item():
     old_name = input('Введите название элемента: ')
     new_name = input('Введите новое название: ')
     if os.path.exists(old_name):
-        if __is_file(old_name):
-            shutil.copy(old_name, new_name)
-        if __is_folder(old_name):
-            shutil.copytree(old_name, new_name)
+        shutil.copy(old_name, new_name) if __is_file(old_name) else shutil.copytree(old_name, new_name)
 
 
 def view_directory():
@@ -42,18 +36,12 @@ def view_directory():
 
 
 def view_folders():
-    folders = []
-    for item in os.listdir():
-        if (__is_folder(item)):
-            folders.append(item)
+    folders = [item for item in os.listdir() if __is_folder(item)]
     print(folders)
 
 
 def view_files():
-    files = []
-    for item in os.listdir():
-        if (__is_file(item)):
-            files.append(item)
+    files = [item for item in os.listdir() if __is_file(item)]
     print(files)
